% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInformationRetrievalMetrics.R
\name{addInformationRetrievalMetrics}
\alias{addInformationRetrievalMetrics}
\title{Compute information retrieval metrics (e.g. accuracy, precision, recall...)}
\usage{
addInformationRetrievalMetrics(data)
}
\arguments{
\item{data}{data frame with four fields (TP, TN, FP, and FN)}
}
\description{
Compute information retrieval metrics (e.g. accuracy, precision, recall...)
}
\seealso{
createConfusionMatrix
}
