% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPredictor.R
\name{createPredictor}
\alias{createPredictor}
\title{Return a closure that can predict the success of a novel}
\usage{
createPredictor(vectorizer, tfidf, model, threshold = 0.5)
}
\arguments{
\item{vectorizer}{a text2vec vectorizer}

\item{tfidf}{tfidf instance used to train model}

\item{model}{a glmnet model}

\item{threshold}{specify the threshold for success}
}
\description{
Return a closure that can predict the success of a novel
}
