% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSplit.R
\name{createSplit}
\alias{createSplit}
\title{Create a train and test split of the novels dataset}
\usage{
createSplit(data, split = 0.8)
}
\arguments{
\item{data}{a data frame of literary texts in "Stonybrook" format}

\item{split}{specify a split ratio (default = .8)}
}
\description{
Create a train and test split of the novels dataset
}
