% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIterator.R
\name{getIterator}
\alias{getIterator}
\title{Get a text2vec iterator over tokenized words}
\usage{
getIterator(text, ids, progressbar = FALSE)
}
\arguments{
\item{text}{a vector of text documents}

\item{ids}{a list of identifiers (e.g. book titles)}

\item{progressbar}{show progress bar if true (default: FALSE)}
}
\description{
Get a text2vec iterator over tokenized words
}
