% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{novels}
\alias{novels}
\title{novels}
\format{A data frame with nine variables: \code{guttenberg_id}, \code{genre}, \code{author},
  \code{response}, \code{title}, \code{file.name}, \code{fold.number}, \code{download.count},
  \code{text}, \code{language} and \code{date}}
\usage{
novels
}
\description{
Training dataset of literary texts based on the article "Success with Style".
}
\keyword{datasets}
