% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results2Text.R
\name{results2Text}
\alias{results2Text}
\title{Convert numeric response to text value}
\usage{
results2Text(response, threshold = 0.5)
}
\arguments{
\item{response}{a value between 0 and 1}

\item{threshold}{specify the threshold for success}
}
\description{
Convert numeric response to text value
}
