% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectByAuthor.R
\name{selectByAuthor}
\alias{selectByAuthor}
\title{Return a logical vector of authors that match the specified pattern}
\usage{
selectByAuthor(author, data)
}
\arguments{
\item{author}{a regex pattern for one or more authors}

\item{data}{a data frame of literary texts in "Stonybrook" format}
}
\description{
Return a logical vector of authors that match the specified pattern
}
