% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMemoryUse.R
\name{showMemoryUse}
\alias{showMemoryUse}
\title{A pretty=printer that shows memory use in current session}
\usage{
showMemoryUse(sort = "size", decreasing = FALSE, limit)
}
\arguments{
\item{sort}{specify sort by "size" or "alphabetical". Default = "size".}

\item{decreasing}{show results in decreasing size order. Default = False.}

\item{limit}{limit listing to specified value. Default = all.}
}
\description{
A pretty=printer that shows memory use in current session
}
