% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2vocab.R
\name{text2vocab}
\alias{text2vocab}
\title{Convert a vector of texts into a vocabulary object}
\usage{
text2vocab(it, documentCount, ngram = c(1L, 3L), documentMinimum = 10,
  stopWords = stopwords())
}
\arguments{
\item{it}{a text2vec token iterator}

\item{documentCount}{count of documents in dataset}

\item{ngram}{a list specifying the min/max for ngrams}

\item{documentMinimum}{a word must appear in no less than n documents}

\item{stopWords}{words to eliminate from the vocabulary object (use character(0) for none)}
}
\description{
Convert a vector of texts into a vocabulary object
}
